fix: move getClassItems from RestController to SasController
