+++ actions/class.RestItems.php
@@ -18,7 +18,6 @@
  */
 
 use oat\tao\model\routing\AnnotationReader\security;
-use oat\taoItems\model\search\ItemClassListService;
 
 /**
  *
@@ -36,18 +35,6 @@ class taoItems_actions_RestItems extends tao_actions_CommonRestModule
         //The service taht implements or inherits get/getAll/getRootClass ... for that particular type of resources
         $this->service = taoItems_models_classes_CrudItemsService::singleton();
     }
-
-    public function getItemClasses()
-    {
-
-           $this->returnJson(
-            $this->getItemClassListService()->getList(
-                $this->getGetParameter('q'),
-                $this->getGetParameter('page')
-            )
-        );
-    }
-
     /**
      * Optionnaly a specific rest controller may declare
      * aliases for parameters used for the rest communication
@@ -73,10 +60,5 @@ class taoItems_actions_RestItems extends tao_actions_CommonRestModule
         */
 
         ];
-    }
-
-    private function getItemClassListService(): ItemClassListService
-    {
-        return $this->getServiceManager()->getContainer()->get(ItemClassListService::class);
     }
 }
